Don't allow delete a master page if we are using it on a layout

+++ modules/apps/layout/layout-page-template-admin-web/src/main/java/com/liferay/layout/page/template/admin/web/internal/portlet/action/DeleteMasterPageMVCActionCommand.java
@@ -15,21 +15,26 @@
 package com.liferay.layout.page.template.admin.web.internal.portlet.action;
 
 import com.liferay.layout.page.template.admin.constants.LayoutPageTemplateAdminPortletKeys;
+import com.liferay.layout.page.template.exception.RequiredLayoutPageTemplateEntryException;
+import com.liferay.layout.page.template.model.LayoutPageTemplateEntry;
 import com.liferay.layout.page.template.service.LayoutPageTemplateEntryService;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.util.ParamUtil;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
-import java.util.ArrayList;
-import java.util.List;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Eudaldo Alonso
@@ -69,23 +74,51 @@ public class DeleteMasterPageMVCActionCommand extends BaseMVCActionCommand {
 		for (long deleteLayoutPageTemplateEntryId :
 				deleteLayoutPageTemplateEntryIds) {
 
-			try {
-				_layoutPageTemplateEntryService.deleteLayoutPageTemplateEntry(
+			LayoutPageTemplateEntry layoutPageTemplateEntry =
+				_layoutPageTemplateEntryService.fetchLayoutPageTemplateEntry(
 					deleteLayoutPageTemplateEntryId);
+
+			if (layoutPageTemplateEntry == null) {
+				SessionErrors.add(actionRequest, PortalException.class);
+
+				deleteLayoutPageTemplateIdsList.add(
+					deleteLayoutPageTemplateEntryId);
+
+				continue;
+			}
+
+			int count = _layoutLocalService.getLayoutsCount(
+				layoutPageTemplateEntry.getGroupId(),
+				layoutPageTemplateEntry.getLayoutPageTemplateEntryId());
+
+			if (count > 0) {
+				SessionErrors.add(
+					actionRequest,
+					RequiredLayoutPageTemplateEntryException.class);
+
+				deleteLayoutPageTemplateIdsList.add(
+					deleteLayoutPageTemplateEntryId);
+
+				continue;
+			}
+
+			try {
+                _layoutPageTemplateEntryService.deleteLayoutPageTemplateEntry(
+                    deleteLayoutPageTemplateEntryId);
 			}
 			catch (PortalException pe) {
 				if (_log.isDebugEnabled()) {
 					_log.debug(pe, pe);
 				}
 
+				SessionErrors.add(actionRequest, PortalException.class);
+
 				deleteLayoutPageTemplateIdsList.add(
 					deleteLayoutPageTemplateEntryId);
 			}
 		}
 
-		if (!deleteLayoutPageTemplateIdsList.isEmpty()) {
-			SessionErrors.add(actionRequest, PortalException.class);
-
+		if (!SessionErrors.isEmpty(actionRequest)) {
 			hideDefaultErrorMessage(actionRequest);
 
 			sendRedirect(actionRequest, actionResponse);
@@ -93,7 +126,10 @@ public class DeleteMasterPageMVCActionCommand extends BaseMVCActionCommand {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		DeleteMasterPageMVCActionCommand.class);
+		DeleteLayoutPageTemplateEntryMVCActionCommand.class);
+
+	@Reference
+	private LayoutLocalService _layoutLocalService;
 
 	@Reference
 	private LayoutPageTemplateEntryService _layoutPageTemplateEntryService;
